moved application init code to correct class
